backend/requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-debug-toolbar==2.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
hiredis==1.1.0
html5lib==1.0.1
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
ipaddr==2.2.0
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==1.0.0
oauthlib==3.1.0
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
progress==1.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
zope.interface==5.1.0
